move `__deep_copy__` back to base model
